중간저장. saleRqustCriteria 에러
