+++ src/main/java/com/ncncn/controller/GifticonController.java
@@ -1,6 +1,6 @@
 package com.ncncn.controller;
 
-import com.ncncn.domain.CriteriaSM;
+import com.ncncn.domain.pagination.CriteriaSM;
 import com.ncncn.domain.PrcUpdateDTO;
 import com.ncncn.domain.UserVO;
 import com.ncncn.service.GifticonService;
@@ -9,7 +9,6 @@ import com.ncncn.util.UserAuthCheckUtils;
 import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
