+++ src/main/java/com/ncncn/controller/ProdController.java
@@ -1,8 +1,7 @@
 package com.ncncn.controller;
 
-import com.ncncn.domain.GiftiCriteria;
-import com.ncncn.domain.PageDTO;
-import com.ncncn.domain.PageDTOHY;
+import com.ncncn.domain.pagination.GiftiCriteria;
+import com.ncncn.domain.pagination.PageDTO;
 import com.ncncn.domain.WishListVO;
 import com.ncncn.service.ProdService;
 import com.ncncn.service.WishListService;
@@ -10,7 +9,6 @@ import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
