--- src/main/java/com/ncncn/domain/Criteria.java
@@ -1,22 +1,0 @@
-package com.ncncn.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class Criteria {
-    private int pageNum;
-    private int amount;
-
-    public Criteria() {
-
-    }
-
-    public Criteria(int pageNum, int amount) {
-        this.pageNum = pageNum;
-        this.amount = amount;
-    }
-}
