+++ src/main/java/com/ncncn/mapper/ProdMapper.java
@@ -1,6 +1,7 @@
 package com.ncncn.mapper;
 
 import com.ncncn.domain.*;
+import com.ncncn.domain.pagination.GiftiCriteria;
 
 import java.util.List;
