+++ src/main/java/com/ncncn/service/WishListServiceImpl.java
@@ -1,6 +1,6 @@
 package com.ncncn.service;
 
-import com.ncncn.domain.CriteriaSM;
+import com.ncncn.domain.pagination.CriteriaSM;
 import com.ncncn.domain.WishListVO;
 import com.ncncn.mapper.WishListMapper;
 import lombok.AllArgsConstructor;
