+++ src/test/java/com/ncncn/mapper/DnSMapperTests.java
@@ -1,6 +1,6 @@
 package com.ncncn.mapper;
 
-import com.ncncn.domain.CriteriaSM;
+import com.ncncn.domain.pagination.CriteriaSM;
 import com.ncncn.domain.MyDealsDTO;
 import com.ncncn.domain.MySellDTO;
 import lombok.Setter;
