+++ src/test/java/com/ncncn/mapper/ProdMapperTests.java
@@ -1,6 +1,7 @@
 package com.ncncn.mapper;
 
 import com.ncncn.domain.*;
+import com.ncncn.domain.pagination.GiftiCriteria;
 import lombok.Setter;
 import lombok.extern.log4j.Log4j;
 import org.junit.Test;
