+++ src/test/java/com/ncncn/mapper/UserCheckMapperTests.java
@@ -1,7 +1,7 @@
 package com.ncncn.mapper;
 
 
-import com.ncncn.domain.CriteriaCH;
+import com.ncncn.domain.pagination.CriteriaCH;
 import com.ncncn.domain.UserDetailCheckDTO;
 import com.ncncn.domain.UserMemoDTO;
 import com.ncncn.domain.UserVO;
