+++ src/test/java/com/ncncn/service/NoticeServiceTests.java
@@ -2,7 +2,7 @@ package com.ncncn.service;
 
 
 
-import com.ncncn.domain.CriteriaCs;
+import com.ncncn.domain.pagination.CriteriaCs;
 import com.ncncn.domain.CsNoticeVO;
 import lombok.Setter;
 import lombok.extern.log4j.Log4j;
