+++ src/test/java/com/ncncn/service/UserCheckServiceTests.java
@@ -1,6 +1,6 @@
 package com.ncncn.service;
 
-import com.ncncn.domain.CriteriaCH;
+import com.ncncn.domain.pagination.CriteriaCH;
 import com.ncncn.domain.UserDetailCheckDTO;
 import com.ncncn.domain.UserMemoDTO;
 import com.ncncn.domain.UserStatusDTO;
